+++ instrumentation/opentelemetry-instrumentation-psycopg2/src/opentelemetry/instrumentation/psycopg2/__init__.py
@@ -227,6 +227,7 @@ class Psycopg2Instrumentor(BaseInstrumentor):
             "enable_attribute_commenter", False
         )
         capture_parameters = kwargs.get("capture_parameters", False)
+        enable_transaction_spans = kwargs.get("enable_transaction_spans", True)
         dbapi.wrap_connect(
             __name__,
             psycopg2,
@@ -240,6 +241,7 @@ class Psycopg2Instrumentor(BaseInstrumentor):
             commenter_options=commenter_options,
             enable_attribute_commenter=enable_attribute_commenter,
             capture_parameters=capture_parameters,
+            enable_transaction_spans=enable_transaction_spans,
         )
 
     def _uninstrument(self, **kwargs):
@@ -320,7 +322,8 @@ class DatabaseApiIntegration(dbapi.DatabaseApiIntegration):
         kwargs["cursor_factory"] = _new_cursor_factory(**new_factory_kwargs)
         connection = connect_method(*args, **kwargs)
         self.get_connection_attributes(connection)
-        return connection
+        # psycopg2 uses cursor_factory for cursor tracing, so disable cursor wrapping
+        return dbapi.get_traced_connection_proxy(connection, self, wrap_cursors=False)
 
 
 class CursorTracer(dbapi.CursorTracer):